# Optimisation for adding milliseconds to timestamps