Signed-off-by: Mark Riddoch <mark@dianomic.com>

Aman, it struck me this morning that given your optimisation for caching the seconds we could; also improve on the sub-second portion by avoiding sprint, that is known to be slow, for both the number conversation and the string concatenation. Also in the FMT_ISO8601MS case doing the conversion to the C++ string is slow.

New verses original times with this code for the abstracted portion
```
original test1: 10000 iterations took 0.002458 seconds
original test2: 10000 iterations took 0.005415 seconds
     new test1: 10000 iterations took 0.001551 seconds
     new test2: 10000 iterations took 0.001933 seconds
```